+++ pylabrobot/utils/list_tests.py
@@ -2,11 +2,7 @@
 
 import unittest
 
-from pylabrobot.utils import (
-  assert_shape,
-  reshape_2d,
-  expand
-)
+from pylabrobot.utils import assert_shape, reshape_2d
 
 
 class TestListUtils(unittest.TestCase):
@@ -29,13 +25,3 @@ class TestListUtils(unittest.TestCase):
       reshape_2d([1, 2, 3, 4], (2, 3))
     with self.assertRaises(ValueError):
       reshape_2d([1, 2, 3, 4, 5, 6], (2, 2))
-
-  def test_expand(self):
-    self.assertEqual(expand(1, n=3), [1, 1, 1])
-    self.assertEqual(expand([1, 2, 3], n=3), [1, 2, 3])
-    self.assertEqual(expand("test", n=3), ["test", "test", "test"])
-
-    with self.assertRaises(ValueError):
-      expand(1, n=0)
-    with self.assertRaises(ValueError):
-      expand(1, n=-1)